backend
This backend allows for the creation of a memory card game which usings
message passing to access the functions within the cardgame. The
cardgame can be made with any number of slots which will be randomly
taken from the currentDeck variable. The algorithm could use some work
for creating the cardgames as it is possible for a cardgame to have to
pairs and the game to be impossible to win. the backend can currently
create a game and check if certain cards in the game are matching return
#t or #f.